refactor: Change file name
* api_params.py -> swagger_schema.py

Resolves: #11
