+++ argus/monitoring/swagger_schema.py
