+++ argus/monitoring/tests/test_script.py
@@ -3,7 +3,7 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 from django.contrib.auth import get_user_model
 from ..models import *
-from ..api_params import *
+from ..swagger_schema import *
 import math
 from typing import Union
 from datetime import datetime
